Changed lat/lng types to Float for location accuracy

+++ app/forms/event_form.py
@@ -1,6 +1,6 @@
 from wsgiref.validate import validator
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, IntegerField
+from wtforms import StringField, SubmitField, IntegerField, FloatField
 from wtforms.validators import DataRequired, ValidationError
 
 
@@ -10,8 +10,8 @@ class EventForm(FlaskForm):
     name = StringField('Name', validators=[DataRequired()])
     address = StringField('Address', validators=[DataRequired()])
     # TODO edit below for google api integration: 
-    lat = IntegerField('Latitude', validators=[DataRequired()])
-    lng = IntegerField('Longitude', validators=[DataRequired()])
+    lat = FloatField('Latitude', validators=[DataRequired()])
+    lng = FloatField('Longitude', validators=[DataRequired()])
     # city = StringField('City', validators=[DataRequired()])
     # state = StringField('State', validators=[DataRequired()])
     # zipcode = StringField('Zip', validators=[DataRequired()])
